Factor out validation functions and tests
